Keep processor capabilities browser-safe
